+++ lib/src/row/table_header_row_build_mixin.dart
@@ -1,7 +1,7 @@
 import 'package:flexible_scrollable_table_view/src/animation/flexible_table_animations.dart';
 import 'package:flexible_scrollable_table_view/src/arguments/table_row_build_arguments.dart';
 import 'package:flexible_scrollable_table_view/src/decoration/flexible_table_decorations.dart';
-import 'package:flexible_scrollable_table_view/src/scrollable/synchronized_scroll_mixin.dart';
+import 'package:flexible_scrollable_table_view/src/scrollable/table_horizontal_scroll_state_widget.dart';
 import 'package:flutter/widgets.dart';
 
 ///表头行构建功能类
@@ -11,7 +11,6 @@ abstract class TableHeaderRowBuildInterface<T> {
     required TableHeaderRowBuildArguments<T> arguments,
     AbsFlexibleTableDecorations<T>? decorations,
     AbsFlexibleTableAnimations<T>? animations,
-    SynchronizedScrollMixin? scrollController,
     ScrollPhysics? physics,
   });
 }
@@ -22,14 +21,12 @@ mixin TableHeaderRowBuildMixin<T> on TableHeaderRowBuildInterface<T> {
     required TableHeaderRowBuildArguments<T> arguments,
     AbsFlexibleTableDecorations<T>? decorations,
     AbsFlexibleTableAnimations<T>? animations,
-    SynchronizedScrollMixin? scrollController,
     ScrollPhysics? physics,
   }) {
     final Widget child = _FlexibleTableHeader<T>(
       arguments,
       decorations: decorations,
       animations: animations,
-      scrollController: scrollController,
       physics: physics,
     );
     final BoxConstraints constraints = arguments.rowConstraint;
@@ -48,14 +45,12 @@ class _FlexibleTableHeader<T> extends StatelessWidget {
     super.key,
     this.decorations,
     this.animations,
-    this.scrollController,
     this.physics,
   });
 
   final TableHeaderRowBuildArguments<T> arguments;
   final AbsFlexibleTableDecorations<T>? decorations;
   final AbsFlexibleTableAnimations<T>? animations;
-  final SynchronizedScrollMixin? scrollController;
   final ScrollPhysics? physics;
 
   @override
@@ -66,16 +61,19 @@ class _FlexibleTableHeader<T> extends StatelessWidget {
       ),
       if (arguments.configurations.scrollableColumns.isNotEmpty)
         Expanded(
-          child: ListView.builder(
-            controller: scrollController ?? arguments.controller.horizontalScrollController,
-            itemCount: arguments.scrollableColumnList.length,
-            scrollDirection: Axis.horizontal,
-            primary: false,
-            padding: EdgeInsets.zero,
-            physics: physics,
-            itemBuilder: (context, index) => arguments.scrollableColumnList[index].buildHeaderCellInternal(
-              arguments,
-              animations,
+          child: TableHorizontalScrollStateWidget(
+            arguments.horizontalScrollMixin,
+            builder: (context, scrollController) => ListView.builder(
+              controller: scrollController,
+              itemCount: arguments.scrollableColumnList.length,
+              scrollDirection: Axis.horizontal,
+              primary: false,
+              padding: EdgeInsets.zero,
+              physics: physics,
+              itemBuilder: (context, index) => arguments.scrollableColumnList[index].buildHeaderCellInternal(
+                arguments,
+                animations,
+              ),
             ),
           ),
         ),